// wow-gym-server/src/models/device-events.ts
export enum DeviceEvent {
    WALKING_START = 'device_walking_start',
    WALKING_STOP = 'device_walking_stop',
    DIRECTION_LEFT = 'device_direction_left',
    DIRECTION_RIGHT = 'device_direction_right',
    DIRECTION_STRAIGHT = 'device_direction_straight',
    TARGET_ENEMY = 'device_target_enemy',
    TARGET_FRIEND = 'device_target_friend',
    TARGET_NONE = 'device_target_none',
}